chore: break down styled components and extract `DebugPanel`
