+++ apps/cowswap-frontend/src/common/pure/OrderProgressBarV2/steps/CancelledStep.tsx
@@ -1,4 +1,4 @@
-import * as styledEl from '../styled'
+import * as styledEl from './styled'
 
 interface CancelledStepProps {
   children: React.ReactNode
